+++ packages/smarthr-ui/src/components/InputFile/InputFileNative.tsx
@@ -3,10 +3,7 @@
 import {
   type ChangeEvent,
   type MouseEvent,
-  type PropsWithChildren,
-  type ReactNode,
   forwardRef,
-  memo,
   useId,
   useImperativeHandle,
   useMemo,
@@ -15,12 +12,11 @@ import {
 } from 'react'
 
 import { useLatest } from '../../hooks/useLatest'
-import { Localizer } from '../../intl'
 import { BaseColumn } from '../Base'
-import { Button } from '../Button'
-import { FaFolderOpenIcon, FaTrashCanIcon } from '../Icon'
 import { Stack } from '../Layout'
 
+import { FilePreviewDialog } from './FilePreviewDialog'
+import { FileListItem, LabelRender, StyledFaFolderOpenIcon } from './parts'
 import { classNameGenerator } from './style'
 
 import type { Props as CommonProps } from './types'
@@ -37,6 +33,7 @@ export const InputFileNative = forwardRef<HTMLInputElement, Props>(
     ref,
   ) => {
     const [files, setFiles] = useState<File[]>([])
+    const [previewFile, setPreviewFile] = useState<File | null>(null)
     const labelId = useId()
 
     const classNames = useMemo(() => {
@@ -61,7 +58,7 @@ export const InputFileNative = forwardRef<HTMLInputElement, Props>(
       () => inputRef.current,
     )
 
-    const latest = useLatest({ onChange, files })
+    const latest = useLatest({ onChange, files, previewFile })
 
     const functions = useMemo(() => {
       const updateFiles = (newFiles: File[]) => {
@@ -95,6 +92,20 @@ export const InputFileNative = forwardRef<HTMLInputElement, Props>(
           inputRef.current.files = buff.files
           isUpdatingFilesRef.current = false
         },
+        handleClosePreview: () => {
+          setPreviewFile(null)
+        },
+        handleDownload: () => {
+          const file = latest.previewFile
+          if (!file) return
+
+          const url = URL.createObjectURL(file)
+          const a = document.createElement('a')
+          a.href = url
+          a.download = file.name
+          a.click()
+          URL.revokeObjectURL(url)
+        },
       }
     }, [latest])
 
@@ -105,12 +116,12 @@ export const InputFileNative = forwardRef<HTMLInputElement, Props>(
             {files.map((file, index) => (
               <FileListItem
                 key={index}
-                value={index}
+                file={file}
+                index={index}
                 handleDeleteClick={functions.handleDelete}
+                handlePreviewClick={setPreviewFile}
                 className={classNames.fileItem}
-              >
-                {file.name}
-              </FileListItem>
+              />
             ))}
           </BaseColumn>
         )}
@@ -129,44 +140,12 @@ export const InputFileNative = forwardRef<HTMLInputElement, Props>(
           <StyledFaFolderOpenIcon className={classNames.prefix} />
           <LabelRender id={labelId} label={label} />
         </span>
+        <FilePreviewDialog
+          file={previewFile}
+          handleClose={functions.handleClosePreview}
+          handleDownload={functions.handleDownload}
+        />
       </Stack>
     )
   },
 )
-
-type FileListItemProps = PropsWithChildren<{
-  value: number
-  handleDeleteClick: (e: MouseEvent<HTMLButtonElement>) => void
-  className: string
-}>
-
-const FileListItem = memo<FileListItemProps>(
-  ({ value, handleDeleteClick, className, children }) => (
-    <li className={className}>
-      <span className="smarthr-ui-InputFile-fileName shr-wrap-break-word shr-min-w-[0]">
-        {children}
-      </span>
-      <Button
-        variant="text"
-        prefix={<FaTrashCanIcon />}
-        value={value}
-        onClick={handleDeleteClick}
-        className="smarthr-ui-InputFile-deleteButton"
-      >
-        <Localizer id="smarthr-ui/InputFile/destroy" defaultText="削除" />
-      </Button>
-    </li>
-  ),
-)
-
-const StyledFaFolderOpenIcon = memo<{ className: string }>(({ className }) => (
-  <span className={className}>
-    <FaFolderOpenIcon />
-  </span>
-))
-
-const LabelRender = memo<{ id: string; label: ReactNode }>(({ id, label }) => (
-  <span id={id} aria-hidden="true">
-    {label}
-  </span>
-))
